move metadata test schemaspace and schemasprovider out of tests dir
