+++ pyproject.toml
@@ -127,16 +127,16 @@ runtimes = "elyra.metadata.schemaspaces:Runtimes"
 runtimes-images = "elyra.metadata.schemaspaces:RuntimeImages"
 code-snippets = "elyra.metadata.schemaspaces:CodeSnippets"
 component-catalogs = "elyra.metadata.schemaspaces:ComponentCatalogs"
-metadata-tests = "elyra.tests.metadata.test_utils:MetadataTestSchemaspace"
+metadata-tests = "elyra.metadata.schemaspaces:MetadataTestSchemaspace"
 
 [project.entry-points."metadata.schemas_providers"]
 runtimes = "elyra.metadata.schemasproviders:RuntimesSchemas"
 runtimes-images = "elyra.metadata.schemasproviders:RuntimeImagesSchemas"
 code-snippets = "elyra.metadata.schemasproviders:CodeSnippetsSchemas"
 component-catalogs = "elyra.metadata.schemasproviders:ComponentCatalogsSchemas"
+metadata-tests = "elyra.metadata.schemasproviders:MetadataTestSchemasProvider"
 airflow-provider-package-catalog-schema = "elyra.pipeline.airflow.provider_package_catalog_connector.airflow_provider_package_schema_provider:AirflowProviderPackageSchemasProvider"
 airflow-package-catalog-schema = "elyra.pipeline.airflow.package_catalog_connector.airflow_package_schema_provider:AirflowPackageSchemasProvider"
-metadata-tests = "elyra.tests.metadata.test_utils:MetadataTestSchemasProvider"
 
 [project.entry-points."elyra.pipeline.processors"]
 local = "elyra.pipeline.local.local_processor:LocalPipelineProcessor"
